+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
   [#10](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/10)
 * GeoJSON exporter
   [#39](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/39)
+* Initial ArcGIS exporter
+  [#11](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/11)
 
 ### Fixed
 
@@ -31,6 +33,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 
 * derived units in latest asset positions export database view
   [#36](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/36)
+* Downgrading to Python 3.11 due to `arcgis` dependency
+  [#11](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/11)
 
 ## [0.2.2] - 2024-04-30